Move ExceptionWithBacktrace to backtrace.hpp
