+++ src/realm/exceptions.hpp
@@ -23,117 +23,10 @@
 
 #include <realm/util/features.h>
 #include <realm/util/backtrace.hpp>
-#include <realm/util/optional.hpp>
 
 namespace realm {
-namespace _impl {
 
-class ExceptionWithBacktraceBase {
-public:
-    ExceptionWithBacktraceBase()
-        : m_backtrace(util::Backtrace::capture())
-    {
-    }
-    const util::Backtrace& backtrace() const noexcept
-    {
-        return m_backtrace;
-    }
-    virtual const char* message() const noexcept = 0;
-
-protected:
-    util::Backtrace m_backtrace;
-    mutable util::Optional<std::string> m_message_with_backtrace;
-
-    // Render the message and the backtrace into m_message_with_backtrace. If an
-    // exception is thrown while rendering the message, the message without the
-    // backtrace will be returned.
-    const char* materialize_message() const noexcept;
-};
-
-} // namespace _impl
-
-
-/// Base class for exceptions that record a stack trace of where they were
-/// thrown.
-///
-/// The template argument is expected to be an exception type conforming to the
-/// standard library exception API (`std::exception` and friends).
-///
-/// It is possible to opt in to exception backtraces in two ways, (a) as part of
-/// the exception type, in which case the backtrace will always be included for
-/// all exceptions of that type, or (b) at the call-site of an opaque exception
-/// type, in which case it is up to the throw-site to decide whether a backtrace
-/// should be included.
-///
-/// Example (a):
-/// ```
-///     class MyException : ExceptionWithBacktrace<std::exception> {
-///     public:
-///         const char* message() const noexcept override
-///         {
-///             return "MyException error message";
-///         }
-///     };
-///
-///     ...
-///
-///     try {
-///         throw MyException{};
-///     }
-///     catch (const MyException& ex) {
-///         // Print the backtrace without the message:
-///         std::cerr << ex.backtrace() << "\n";
-///         // Print the exception message and the backtrace:
-///         std::cerr << ex.what() << "\n";
-///         // Print the exception message without the backtrace:
-///         std::cerr << ex.message() << "\n";
-///     }
-/// ```
-///
-/// Example (b):
-/// ```
-///     class MyException : std::exception {
-///     public:
-///         const char* what() const noexcept override
-///         {
-///             return "MyException error message";
-///         }
-///     };
-///
-///     ...
-///
-///     try {
-///         throw ExceptionWithBacktrace<MyException>{};
-///     }
-///     catch (const MyException& ex) {
-///         // Print the exception message and the backtrace:
-///         std::cerr << ex.what() << "\n";
-///     }
-/// ```
-template <class Base = std::runtime_error>
-class ExceptionWithBacktrace : public Base, public _impl::ExceptionWithBacktraceBase {
-public:
-    template <class... Args>
-    inline ExceptionWithBacktrace(Args&&... args)
-        : Base(std::forward<Args>(args)...)
-        , _impl::ExceptionWithBacktraceBase() // backtrace captured here
-    {
-    }
-
-    /// Return the message of the exception, including the backtrace of where
-    /// the exception was thrown.
-    const char* what() const noexcept final
-    {
-        return materialize_message();
-    }
-
-    /// Return the message of the exception without the backtrace. The default
-    /// implementation calls `Base::what()`.
-    const char* message() const noexcept override
-    {
-        return Base::what();
-    }
-};
+using util::ExceptionWithBacktrace;
 
 /// Thrown by various functions to indicate that a specified table does not
 /// exist.
